Separate hooks to its own file
