+++ src/Nethermind/Nethermind.Merge.Plugin/Synchronization/MergeBlockDownloader.cs
@@ -2,8 +2,6 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System;
-using System.Diagnostics;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Nethermind.Blockchain;
@@ -17,7 +15,6 @@ using Nethermind.Core.Extensions;
 using Nethermind.Core.Specs;
 using Nethermind.Crypto;
 using Nethermind.Logging;
-using Nethermind.Network;
 using Nethermind.Synchronization;
 using Nethermind.Synchronization.Blocks;
 using Nethermind.Synchronization.ParallelSync;
@@ -170,42 +167,6 @@ namespace Nethermind.Merge.Plugin.Synchronization
                 // already processed header starting from `ProcessDestination`.
                 _beaconPivot.ProcessDestination = currentBlock.Header;
             }
-        }
-    }
-
-    public class PosTransitionHook(IBlockTree blockTree, IPoSSwitcher poSSwitcher, ILogManager logManager) : IPosTransitionHook
-    {
-        private readonly ILogger _logger = logManager.GetClassLogger<PosTransitionHook>();
-
-        public void TryUpdateTerminalBlock(BlockHeader currentHeader, bool shouldProcess)
-        {
-            if (shouldProcess == false) // if we're processing the block we will find TerminalBlock after processing
-                poSSwitcher.TryUpdateTerminalBlock(currentHeader);
-        }
-
-        public bool ImprovementRequirementSatisfied(PeerInfo? bestPeer)
-        {
-            return bestPeer!.TotalDifficulty > (blockTree.BestSuggestedHeader?.TotalDifficulty ?? 0) &&
-                   poSSwitcher.HasEverReachedTerminalBlock() == false;
-        }
-
-        public IOwnedReadOnlyList<BlockHeader> FilterPosHeader(IOwnedReadOnlyList<BlockHeader> response)
-        {
-            // Override PoW's RequestHeaders so that it won't request beyond PoW.
-            // This fixes `Incremental Sync` hive test.
-            if (response.Count > 0)
-            {
-                BlockHeader lastBlockHeader = response[^1];
-                bool lastBlockIsPostMerge = poSSwitcher.GetBlockConsensusInfo(response[^1]).IsPostMerge;
-                if (lastBlockIsPostMerge) // Initial check to prevent creating new array every time
-                {
-                    response = response
-                        .TakeWhile(header => !poSSwitcher.GetBlockConsensusInfo(header).IsPostMerge)
-                        .ToPooledList(response.Count);
-                    if (_logger.IsInfo) _logger.Info($"Last block is post merge. {lastBlockHeader.Hash}. Trimming to {response.Count} sized batch.");
-                }
-            }
-            return response;
         }
     }
 }
